textfinder: add recipe
Add a recipe for the 'textfinder' (copied) sample qt5 application. The code
for this sample application can be found at:

	https://github.com/twoerner/textfinder

This recipe provides an example of building a qt5 application using meta-qt5
from the current master. To try this example you'll need to add the following
layers to conf/bblayers.conf:

	meta-openembedded/meta-oe
	meta-openembedded/meta-multimedia
	meta-openembedded/meta-ruby
	meta-qt5
	meta-trevor

Then in your conf/local.conf add something along the lines of:

	IMAGE_INSTALL_append = " qtbase qtbase-plugins textfinder"

And build with:

	$ bitbake core-image-x11

NOTE: as master is always changing I can't know for how long this recipe will
      continue to work, it works today...

Signed-off-by: Trevor Woerner <twoerner@gmail.com>