layer to misc cleanup
This had been intended to be an OE layer, but over time it wasn't used so much
as a layer as it was used as a general dumping grounds for notes, scripts,
configurations, etc. This commit officially changes this repository from
pretending to be a layer, to conveying its actual purpose.

It had been called "meta-trevor" but is now "oe-misc".

Signed-off-by: Trevor Woerner <twoerner@gmail.com>