gitcmd.sh: add "all up-to-date" status
Have the script keep track of whether or not each repository was already
up-to-date and if none of them needed to be updated, print the line "All
repositories are up-to-date" and return a special return status of 255.

This is useful for, say, Jenkins. If a Jenkins job pulls all the relevant
repositories before performing a build but can determine that all the
repositories are already up-to-date, then it can circumvent the build avoiding
a potentially long job run that would otherwise be identical to the previous.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>