gitcmd.sh: add post-msg
Sometimes a pull causes a lot of activity and I'd like to see which repository
caused such activity. But if the activity scrolls off the top of the window,
it isn't easy to see what repository is responsible. Therefore print which
repository was just done before moving on to the next one.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>