gitcmd.sh: add KEEPGOING
Add handling of cmdline args. Add an optional cmdline arg to allow the gitcmd
to keep going in the face of errors.

Previously all errors were simply ignored. Now errors will be reported. The
gitcmd will stop immediately when it encounters an erroneous return from the
git command, unless the KEEPGOING flag is in force (-k|--keep-going), in which
case it will continue working, and report the error(s) at the end.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>